// store/actions/types/reports.ts
//reports
export enum GetReportsTypes {
  GET_REPORTS_LOADING = "GET_REPORTS_LOADING",
  GET_REPORTS_SUCCESS = "GET_REPORTS_SUCCESS",
  GET_REPORTS_ERROR = "GET_REPORTS_ERROR",
}

export enum GetRecentReportsTypes {
  GET_RECENT_REPORTS_LOADING = "GET_RECENT_REPORTS_LOADING",
  GET_RECENT_REPORTS_SUCCESS = "GET_RECENT_REPORTS_SUCCESS",
  GET_RECENT_REPORTS_ERROR = "GET_RECENT_REPORTS_ERROR",
}

//reasons
export enum GetReportsReasonsTypes {
  GET_REPORTS_REASONS_LOADING = "GET_REPORTS_REASONS_LOADING",
  GET_REPORTS_REASONS_SUCCESS = "GET_REPORTS_REASONS_SUCCESS",
  GET_REPORTS_REASONS_ERROR = "GET_REPORTS_Reasons_ERROR",
}

export enum ReportsReasonsOperationsTypes {
  OPERATION_REPORTS_REASONS_LOADING = "OPERATION_REPORTS_REASONS_LOADING",
  OPERATION_REPORTS_REASONS_SUCCESS = "OPERATION_REPORTS_REASONS_SUCCESS",
  OPERATION_REPORTS_REASONS_ERROR = "OPERATION_REPORTS_REASONS_ERROR",
}